Update components folder path
